Bugfix bc I did not check everything correctly

+++ src_main/tools/file_operations.py
@@ -36,11 +36,12 @@ def remove_sim_instance_folders(data_path, uids):
 
 
 def remove_design_point_configurations_dummy_path(sim_dummy_directory, pattern="custom_dummy_*"):
-    simulation_run_dir_pattern = os.path.join(sim_dummy_directory, pattern)
-    matching_sim_dummy_dir = [path for path in glob.glob(simulation_run_dir_pattern) if os.path.isdir(path)]
-    for directory in matching_sim_dummy_dir:
-        if os.path.exists(directory):
-            shutil.rmtree(directory)
+    if sim_dummy_directory is not None and os.path.exists(sim_dummy_directory):
+        simulation_run_dir_pattern = os.path.join(sim_dummy_directory, pattern)
+        matching_sim_dummy_dir = [path for path in glob.glob(simulation_run_dir_pattern) if os.path.isdir(path)]
+        for directory in matching_sim_dummy_dir:
+            if os.path.exists(directory):
+                shutil.rmtree(directory)
 
 
 def remove_design_point_configurations_sims_path(sims_path):
